Add --non-interactive switch to focker bootstrap.

+++ focker/bootstrap.py
@@ -7,6 +7,8 @@ from .zfs import zfs_poolname, \
 
 
 def command_bootstrap(args):
+    if args.empty and args.non_interactive:
+        raise ValueError('--empty and --non-interactive are mutually exclusive')
     version = subprocess.check_output(['freebsd-version']).decode('utf-8')
     print('FreeBSD version:', version)
     tags = args.tags or [ 'freebsd-' + version.split('-')[0], 'freebsd-latest' ]
@@ -15,10 +17,13 @@ def command_bootstrap(args):
     name = find_prefix(poolname + '/focker/images/', sha256)
     subprocess.check_output(['zfs', 'create', '-o', 'focker:sha256=' + sha256, name])
     zfs_tag(name, tags)
-    if not args.empty:
+    res = None
+    if args.non_interactive:
+        res = subprocess.run(['focker-bsdinstall', zfs_mountpoint(name)])
+    elif not args.empty:
         res = subprocess.run(['bsdinstall', 'jail', zfs_mountpoint(name)])
-        if res.returncode != 0:
-            zfs_run(['zfs', 'destroy', '-r', '-f', name])
-            raise ValueError('bsdinstall failed')
+    if res is not None and res.returncode != 0:
+        zfs_run(['zfs', 'destroy', '-r', '-f', name])
+        raise ValueError('bsdinstall failed')
     subprocess.check_output(['zfs', 'set', 'rdonly=on', name])
     subprocess.check_output(['zfs', 'snapshot', name + '@1'])
